refactor(reports): extract standup report to reports_standup.go
